+++ scout/server/blueprints/phenomodels/views.py
@@ -18,7 +18,7 @@ phenomodels_bp = Blueprint(
 
 
 @phenomodels_bp.route("/<institute_id>/advanced_phenotypes", methods=["GET"])
-@templated("phenomodels.html")
+@templated("phenomodels/phenomodels.html")
 def advanced_phenotypes(institute_id):
     """Show institute-level advanced phenotypes"""
     institute_obj = institute_and_case(store, institute_id)
@@ -101,7 +101,7 @@ def phenomodel_edit(institute_id, model_id):
 
 
 @phenomodels_bp.route("/<institute_id>/phenomodel/<model_id>", methods=["GET"])
-@templated("phenomodel.html")
+@templated("phenomodels/phenomodel.html")
 def phenomodel(institute_id, model_id):
     """View/Edit an advanced phenotype model"""
     institute_obj = institute_and_case(store, institute_id)
@@ -125,3 +125,40 @@ def phenomodel(institute_id, model_id):
         phenomodel=phenomodel_obj,
         subpanel_form=subpanel_form,
     )
+
+
+@phenomodels_bp.route("/<institute_id>/phenomodel_assign/<case_name>", methods=["POST"])
+def phenomodel_assign(institute_id: str, case_name: str):
+    """Assign multiple phenotypes from an advanced model to one or more individuals/samples of a case."""
+    institute_obj, case_obj = institute_and_case(store, institute_id, case_name)
+    case_url = url_for("cases.case", institute_id=institute_id, case_name=case_name)
+    user_obj = store.user(current_user.email)
+
+    if not case_obj:
+        return safe_redirect_back(request)
+
+    individuals = request.form.getlist("phenotype_inds")
+
+    for pheno in request.form.getlist("checked_terms"):
+        term = pheno.split(".")[-1]
+        if term.startswith("HP:"):
+            case_obj = store.add_phenotype(
+                institute=institute_obj,
+                case=case_obj,
+                user=user_obj,
+                link=case_url,
+                hpo_term=term,
+                phenotype_inds=individuals,
+            )
+        else:  # OMIM, ORPHA
+            case_obj = store.diagnose(
+                institute=institute_obj,
+                case=case_obj,
+                user=user_obj,
+                link=case_url,
+                disease_id=term,
+                affected_inds=individuals,
+                remove=False,
+            )
+
+    return safe_redirect_back(request)