+++ CHANGELOG.md
@@ -11,6 +11,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Accept single end BNDs (#6132)
 - Update analysis type for case's individual/sample using `scout update individual` command (#6142)
 - Filter methbat outliers by significance ("HypoMethylated", "HyperMethylated", "HypoASM", "HyperASM") (#6140)
+- Pick phenotype terms (HPO, OMIM, ORPHA) to be assigned to a case from a phenomodel object (#6138)
 ### Changed
 - Export genotype with gene variantS search (#6088)
 - igv.js version to 3.8.0 - fixes `e.tags` error popup issue (#6096, #6118)